apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yasar-nuri-api"
version = "0.1.0"
description = "YasarNuri FastAPI backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "sqlalchemy[asyncio]>=2.0.36",
  "asyncpg>=0.30.0",
  "alembic>=1.14.0",
  "pgvector>=0.3.6",
  "redis>=5.2.0",
  "pydantic-settings>=2.6.0",
  "httpx>=0.28.0",
  "structlog>=24.4.0",
  "google-generativeai>=0.8.3",
  "tenacity>=9.0.0",
  "langdetect>=1.0.9",
  "python-multipart>=0.0.17",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "httpx>=0.28.0",
  "ruff>=0.8.0",
  "mypy>=1.13.0",
  "types-redis",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["yasar_nuri_api*"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.12"
strict = false
packages = ["yasar_nuri_api"]
ignore_missing_imports = true
warn_return_any = false
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
